Rounding improvement & version up
Eyes are now activated on round-offs of points (using native RF `otRound`). This allows eyes to appear during transformations with the transform tool before `mouseUp`. Another, perhaps debatable side-effect is the eye appearing on nearby guides even if contours are set not to snape to individual units in preferences. This covers off on #2